parallel API impl
