+++ src/test/java/baseEntities/BaseApiTest.java
@@ -1,10 +1,11 @@
 package baseEntities;
 
+import com.google.gson.Gson;
 import configuration.ReadProperties;
 import io.restassured.RestAssured;
 import io.restassured.http.ContentType;
 import org.apache.http.protocol.HTTP;
-import org.testng.annotations.BeforeTest;
+import org.testng.annotations.BeforeClass;
 import services.ProjectService;
 import services.TestCaseService;
 
@@ -14,9 +15,13 @@ public class BaseApiTest {
 
     protected TestCaseService testCaseService;
     protected ProjectService projectService;
+    protected Gson gson;
 
-    @BeforeTest
+
+    @BeforeClass
     public void setupApi() {
+        gson = new Gson();
+
         testCaseService = new TestCaseService();
         projectService = new ProjectService();
         RestAssured.baseURI = ReadProperties.getBaseApiUrl();
