--- src/test/resources/dataApiTest/projectForApiTest.json
@@ -1,3 +1,0 @@
-{
-  "name": "Project By API"
-}
