+++ src/test/resources/parallelAPI.xml
@@ -1,0 +1,11 @@
+<?xml version="1.0" encoding="UTF-8" ?>
+<!DOCTYPE suite SYSTEM "http://testng.org/testng-1.0.dtd">
+<suite name="API Suite">
+    <test name="testCRUD" parallel="classes" thread-count="2">
+        <classes>
+            <class name="tests.api.ApiCaseTest"/>
+            <class name="tests.api.ApiProjectTest"/>
+        </classes>
+
+    </test>
+</suite>
